+++ CopilotKit/packages/react-core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # ui
 
+## 0.8.0-alpha.3
+
+### Minor Changes
+
+- textarea edits mvp
+
+## 0.8.0-alpha.2
+
+### Patch Changes
+
+- rebase master
+
 ## 0.8.0-alpha.1
 
 ### Patch Changes