+++ CopilotKit/packages/react-textarea/CHANGELOG.md
@@ -1,5 +1,24 @@
 # ui
 
+## 0.19.0-alpha.3
+
+### Minor Changes
+
+- textarea edits mvp
+
+### Patch Changes
+
+- Updated dependencies
+  - @copilotkit/react-core@0.8.0-alpha.3
+
+## 0.19.0-alpha.2
+
+### Patch Changes
+
+- rebase master
+- Updated dependencies
+  - @copilotkit/react-core@0.8.0-alpha.2
+
 ## 0.19.0-alpha.1
 
 ### Patch Changes